client: Store result from XGetWMHints, track scaled icons
We're supposed to XFree() the result of XGetWMHints but we need it
to know a client's icon pixmap if we need to redraw it, so keep it
around and XFree() it just before we get new hints.

Also track whether we drew our own scaled pixmaps so we can properly
XFreePixmap() them before changing them to something else or
destroying the client.