*: Add configurable desktop mouse bindings
This keeps the default of right click bringing up the launcher by
default, but also adds mouse wheel buttons (4 and 5) to switch to
the next and previous desktops by default.

A "launcher" action is added to support this, which means it can
also be brought up from a keyboard binding so make launcher_show()
handle this properly.