launcher: Update x/y after mapping, don't highlight by default
If we clicked in the corner of the screen, mapping will move the
launcher window to fit on the screen so update x/y.

Also don't highlight by default because our cursor won't necessarily
be at the first entry now, but also because accidentally
right-clicking on the root shouldn't just launch the first thing on
the list.